+++ subsystems/posix/userland/progman_registry.c
@@ -39,6 +39,7 @@ struct progman_default_item
 static const struct progman_default_item k_default_items[] = {
     {"Main", "Shell", "/bin/shellapp", "Terminal and builtins", DESKTOP_ICON_SHELL, SAVANXP_DESKTOP_LAUNCH_FLAG_NONE},
     {"Main", "Files", "/bin/filesapp", "Browse /disk and preview files", DESKTOP_ICON_DESKTOP, SAVANXP_DESKTOP_LAUNCH_FLAG_NONE},
+    {"Main", "Notepad", "/bin/notepad", "Edit text files", DESKTOP_ICON_DESKTOP, SAVANXP_DESKTOP_LAUNCH_FLAG_NONE},
     {"Main", "About", "/bin/aboutapp", "System overview and help", DESKTOP_ICON_DESKTOP, SAVANXP_DESKTOP_LAUNCH_FLAG_NONE},
     {"Games", "Doom", "/disk/bin/doomgeneric", "Classic FPS test port", DESKTOP_ICON_DOOM, SAVANXP_DESKTOP_LAUNCH_FLAG_FULLSCREEN},
 #if DESKTOP_INCLUDE_TEST_APPS